refactor(core): restructure core utils and state management
