+++ packages/muffincss-core/src/state/resolved-classlist-collector.ts
