+++ packages/muffincss-core/src/utils/cleanup-root.ts
